+++ R/pseudo_r2.R
@@ -29,6 +29,11 @@
 #' @export
 pseudo_r2 <- function(x) {
   .Deprecated("r2", package = "sjmisc", msg = "`pseudo_r2` will be deprecated in future versions of `sjmisc`. Please use `r2` instead.")
+  pseudo_ralt(x)
+}
+
+
+pseudo_ralt <- function(x) {
   n <- nrow(x$model)
   CoxSnell <- 1 - exp((x$deviance - x$null) / n)
   Nagelkerke <- CoxSnell / (1 - exp(-x$null / n))
@@ -36,7 +41,6 @@ pseudo_r2 <- function(x) {
   names(Nagelkerke) <- "Nagelkerke"
   return(structure(class = "sjmisc_r2", list(CoxSnell = CoxSnell, Nagelkerke = Nagelkerke)))
 }
-
 
 #' @title Tjur's Coefficient of Discrimination
 #' @name cod
@@ -144,7 +148,7 @@ cod <- function(x) {
 #' @references \itemize{
 #'               \item \href{http://glmm.wikidot.com/faq}{DRAFT r-sig-mixed-models FAQ}
 #'               \item Byrnes, J. 2008. Re: Coefficient of determination (R^2) when using lme(). \href{http://thread.gmane.org/gmane.comp.lang.r.lme4.devel/684}{gmane.comp.lang.r.lme4.devel}
-#'               \item Xu, R. 2003. Measuring explained variation in linear mixed effects models. Statist. Med. 22:3527-3541. \url{doi:10.1002/sim.1572}
+#'               \item Xu, R. 2003. Measuring explained variation in linear mixed effects models. Statist. Med. 22:3527-3541. \doi{10.1002/sim.1572}
 #'               \item Tjur T. 2009. Coefficients of determination in logistic regression models - a new proposal: The coefficient of discrimination. The American Statistician, 63(4): 366-372
 #'             }
 #'
@@ -173,7 +177,7 @@ r2 <- function(x) {
   adjr2 <- NULL
   # do we have a glm? if so, report pseudo_r2
   if (any(class(x) == "glm")) {
-    return(pseudo_r2(x))
+    return(pseudo_ralt(x))
     # do we have a glmer?
   } else if (any(class(x) == "glmerMod")) {
     return(cod(x))