+++ R/rec_pattern.R
@@ -10,11 +10,11 @@
 #' @param width Numeric, indicating the range of each group.
 #' @param other String token, indicating how to deal with all other values
 #'          that have not been captured by the recode pattern. See 'Details'
-#'          on the \code{emph}-token in \code{\link{rec}}.
+#'          on the \code{else}-token in \code{\link{rec}}.
 #' @return A list with two values:
-#'           \itemize{
-#'            \item{pattern} string pattern that can be used as \code{recodes} argument for the \code{\link{rec}}-function.
-#'            \item{labels} the associated values labels that can be used with \code{\link{set_labels}}.
+#'           \describe{
+#'            \item{\code{pattern}}{string pattern that can be used as \code{recodes} argument for the \code{\link{rec}}-function.}
+#'            \item{\code{labels}}{the associated values labels that can be used with \code{\link{set_labels}}.}
 #'           }
 #'
 #' @seealso \code{\link{group_var}} for recoding variables into smaller groups, and